Move calculateCustomSubscriptionForNeighbour to NeighbourService
